Added a generic filter container
This is a component that should be used for filtering content. Appears as a row of selectable rainbow buttons.

Can be used as a checkbox or radio mechanism.

No state is held within the container - state should be managed elsewhere.

Started with basing off the blog filtering section, but modified the API and logic to better fit a generic use case.