Visitor Information page (#577)
* Added a generic filter container

This is a component that should be used for filtering content. Appears as a row of selectable rainbow buttons.

Can be used as a checkbox or radio mechanism.

No state is held within the container - state should be managed elsewhere.

Started with basing off the blog filtering section, but modified the API and logic to better fit a generic use case.

* Moved filterTypes to constants

This is required in another module outside of the original one, so it makes sense to centralise it.

* Tidied up component returns

Return straight from the arrow function, don't bother with the explicit return since there's no logic within the component.

* Added a generic paged card container component

This component can be used to show a gallery of cards in a paged fashion. There is a "show more" button that can expand the currently displayed number of cards by the page size defined.

* Updated imports of PropTypes to be capitalised

* create pageintro block component

* add page intro component to sponsors page, add community partners graphql query to vi page

* community partner card component wip

* community partner card component done

* Extended filterContainer to take hexColour

This is going to be required in order to ensure card tags are consistent with filter colours.

* Moved state out of paged card container component

* Added "Everything" to partners categories

* Pass filter name into onClick handler

* Regenerated snapshots for paged container tests

* Added a filtered, paged, card container

This is with state and logic. There's a bug in it somewhere which means the filtering only works for "everything" right now. Will pick this up tomorrow.

* Regenerated failing snapshots

* Moved onClick to Input

This was the cause of the double-execution bug.

* Cleaned up React errors

* Fixed bug for when a non-all category is deselected

* Added filter paged card container to visitor info page

* Updated filtered paged card container state

Container will now snap to one page on filter change.

* Added ids to things

* Added aria label to filter label

* Renamed categories array

* Updated tests to reduce code duplication

* Reduced dupe in gender icon and community card

* Added key to social map

* Conditionally render accessibility details

* Reduced code dupe in map pin icon

* Reduced complexity in icons

* Further reduced code dupe in tests

* Fixed test logic which wasn't quite right

* Pulled social bar down in desktop view

For the community partner card.

* Used flex to pull social icons down

This avoids a deadzone where icons can float over the top of the gender neutral toilets marker.

* Added cond to not render container if no partners

* Updated filter container

Don't show the filter title if no cards appear for that category.

* Removed unused component from import

* Updated padding in filter label to match design

* Updated test snapshots for filter padding update

* Reduced padding in partner card category tags

* Removed margin from community partenr card item

* Added padding to bottom of fully loaded container

When the cards were fully loaded in the paged card container, there was no padding between the bottom of the container and the footer.

Conditionally add the padding if the container is fully loaded.

* Updated partner card padding for mobile view

* Gatsby image banners (#638)

* add logic to handle gatsby img in headers. Update generic page img qraphql query to get responsive imgs

* mock ie11 polyfill component with regular one to get past import error in jest

* update snapshot, fix package json to match master

* add aria hidden attribute to banner img element

* make event detail page banners use gatsby img

* remove console.log

* update responsive banners for events and blog details

* Add Intercom on all pages (#639)

* Parade map (#616)

* Parade map

* Parade map styles file fix, download PDF button hide

* Map on mobile height fix, badges accesslint issues fix

* links fixes

* parade map appearance changed

* Add parade map link in menu

* Update parade map

* Revert "Merge branch 'issue/567-google-parade-map' of https://github.com/PrideInLondon/pride-london-web into issue/567-google-parade-map"

This reverts commit ae601c6869d519a489589a77c5832712a13297be, reversing
changes made to a58695e192fde4acd7cc02588bc12b00b9f905a9.

* Update parade map

* Update parade map

* Update parade map

* footer logos IE fix (#659)

* redirect added, navigation changed, explore-london copy changed

* copy changed

* navigation padding changed