WEBREF-49 :package: Resolved react-side-effect dependency
This dep was previously resolving to a version which was actively using a deprecated lifecycle method and causing noise in the build/test log.

The makers of the react-helmet library (which uses the dep), are working on a new release to overcome this: https://github.com/nfl/react-helmet/issues/413#issuecomment-578887682.

We'll pick this up with dependabot once it's out of beta.